requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.0=py38_0
asgiref=3.2.7=py_0
attrs=19.3.0=py_0
autobahn=20.4.3=pypi_0
automat=20.2.0=pypi_0
backcall=0.1.0=py38_0
blas=1.0=mkl
bleach=3.1.4=py_0
bson=0.5.8=pypi_0
ca-certificates=2020.4.5.1=hecc5488_0
certifi=2020.4.5.1=py38h32f6830_0
cffi=1.14.0=pypi_0
channels=2.4.0=pypi_0
constantly=15.1.0=pypi_0
cryptography=2.9.2=pypi_0
cyrus-sasl=2.1.27=hce6a660_1
daphne=2.5.0=pypi_0
dataclasses=0.6=pypi_0
dbus=1.13.14=h517e14e_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
django=3.0.3=py_0
django-filter=2.0.0=py_0
djangorestframework=3.11.0=py_0
djongo=1.3.2=pypi_0
entrypoints=0.3=py38_0
eventkit=0.8.6=pypi_0
expat=2.2.6=h0a44026_0
gettext=0.19.8.1=hb0f4f8b_2
glib=2.63.1=h70d4741_1
hyperlink=19.0.0=pypi_0
ib-insync=0.9.60=pypi_0
icu=58.2=h0a44026_3
idna=2.9=pypi_0
importlib_metadata=1.5.0=py38_0
incremental=17.5.0=pypi_0
intel-openmp=2019.4=233
ipykernel=5.1.4=py38h39e3cac_0
ipython=7.13.0=py38h5ca1d4c_0
ipython_genutils=0.2.0=py38_0
ipywidgets=7.5.1=py_0
jedi=0.17.0=py38_0
jinja2=2.11.2=py_0
jpeg=9b=he5867d9_2
jsonschema=3.2.0=py38_0
jupyter=1.0.0=py38_7
jupyter_client=6.1.3=py_0
jupyter_console=6.1.0=py_0
jupyter_core=4.6.3=py38_0
krb5=1.17.1=hddcf347_0
libcxx=10.0.0=1
libedit=3.1.20181209=hb402a30_0
libffi=3.3=h0a44026_1
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libntlm=1.4=h0b31af3_1002
libpng=1.6.37=ha441bb4_0
librdkafka=1.4.0=hfe55218_0
libsodium=1.0.16=h3efe00b_0
lz4-c=1.9.2=h4a8c4bd_1
markupsafe=1.1.1=py38h1de35cc_0
mistune=0.8.4=py38h1de35cc_1000
mkl=2019.4=233
mkl-service=2.3.0=py38hfbe908c_0
mkl_fft=1.0.15=py38h5e564d8_0
mkl_random=1.1.0=py38h6440ff4_0
mongo=0.2.0=pypi_0
nbconvert=5.6.1=py38_0
nbformat=5.0.6=py_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.3.3=pypi_0
notebook=6.0.3=py38_0
numpy=1.18.4=pypi_0
numpy-base=1.18.1=py38h3304bdc_1
openssl=1.1.1g=h0b31af3_0
pandas=1.0.3=py38h6c726b0_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py38_1
parso=0.7.0=py_0
pcre=8.43=h0a44026_0
pexpect=4.8.0=py38_0
pickleshare=0.7.5=py38_1000
pip=20.0.2=py38_3
prometheus_client=0.7.1=py_0
prompt-toolkit=3.0.4=py_0
prompt_toolkit=3.0.4=0
ptyprocess=0.6.0=py38_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pypi_0
pygments=2.6.1=py_0
pyhamcrest=2.0.2=pypi_0
pymongo=3.10.1=pypi_0
pyopenssl=19.1.0=pypi_0
pyqt=5.9.2=py38h655552a_2
pyrsistent=0.16.0=py38h1de35cc_0
python=3.8.3=h26836e1_1
python-confluent-kafka=1.3.0=py38h64e0658_1
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytz=2020.1=py_0
pyzmq=18.1.1=py38h0a44026_0
qt=5.9.7=h468cd18_1
qtconsole=4.7.4=py_0
qtpy=1.9.0=py_0
readline=8.0=h1de35cc_0
send2trash=1.5.0=py38_0
service-identity=18.1.0=pypi_0
setuptools=46.4.0=py38_0
sip=4.19.8=py38h0a44026_0
six=1.14.0=py38_0
sqlite=3.31.1=h5c1f38d_1
sqlparse=0.2.4=pypi_0
terminado=0.8.3=py38_0
testpath=0.4.4=py_0
tk=8.6.8=ha441bb4_0
tornado=6.0.4=py38h1de35cc_1
traitlets=4.3.3=py38_0
twisted=20.3.0=pypi_0
txaio=20.4.1=pypi_0
wcwidth=0.1.9=py_0
webencodings=0.5.1=py38_1
wheel=0.34.2=py38_0
widgetsnbextension=3.5.1=py38_0
xz=5.2.5=h1de35cc_0
zeromq=4.3.1=h0a44026_3
zipp=3.1.0=py_0
zlib=1.2.11=h1de35cc_3
zope-interface=5.1.0=pypi_0
